# Reduce overhead of function call and deprecate rarely used utilities