--- a/pytensor/graph/null_type.py
+++ b/pytensor/graph/null_type.py
@@ -26,9 +26,6 @@ class NullType(Type):
     def filter_variable(self, other, allow_convert=True):
         raise ValueError("No values may be assigned to a NullType")
 
-    def may_share_memory(a, b):
-        return False
-
     def values_eq(self, a, b, force_same_dtype=True):
         raise ValueError("NullType has no values to compare")