--- a/pytensor/tensor/random/op.py
+++ b/pytensor/tensor/random/op.py
@@ -387,24 +387,17 @@ class RandomVariable(Op):
         return node.inputs[2:]
 
     def perform(self, node, inputs, outputs):
-        rng_var_out, smpl_out = outputs
-
         rng, size, *args = inputs
 
         # Draw from `rng` if `self.inplace` is `True`, and from a copy of `rng` otherwise.
         if not self.inplace:
             rng = copy(rng)
 
-        rng_var_out[0] = rng
-
-        if size is not None:
-            size = tuple(size)
-        smpl_val = self.rng_fn(rng, *([*args, size]))
-
-        if not isinstance(smpl_val, np.ndarray) or str(smpl_val.dtype) != self.dtype:
-            smpl_val = np.asarray(smpl_val, dtype=self.dtype)
-
-        smpl_out[0] = smpl_val
+        outputs[0][0] = rng
+        outputs[1][0] = np.asarray(
+            self.rng_fn(rng, *args, None if size is None else tuple(size)),
+            dtype=self.dtype,
+        )
 
     def grad(self, inputs, outputs):
         return [