@ricardoV94 2024-10-22
Marking as a draft because I want to add a note on the docstrings of pytensor.In about mutable and alias